Update Specs to Support Multiple Results
