+++ Wemotely/Support/Platform.swift
@@ -1,3 +1,5 @@
+import UIKit
+
 // http://themainthread.com/blog/2015/06/simulator-check-in-swift.html
 struct Platform {
     static let isSimulator: Bool = {
@@ -7,4 +9,8 @@ struct Platform {
         #endif
         return isSim
     }()
+
+    static func isRunningTests() -> Bool {
+        return CommandLine.arguments.contains("--uitesting") || (NSClassFromString("XCTest") != nil)
+    }
 }
