+++ WemotelyUITests/SegueUITests.swift
@@ -112,22 +112,25 @@ class SegueUITests: XCTestCase {
     }
 
     func testNavigatingToJobs() {
+        let accountName = "All Inboxes"
+
         app.launch()
 
         app.runWithSupportedOrientations {
             startAndEndOnDashboard {
                 // Tap first cell in 'Dashboard' table
-                app.tables["dashboardTableView"].cells.staticTexts["All Inboxes"].tap()
+                app.tables["dashboardTableView"].cells.staticTexts[accountName].tap()
 
                 XCTAssert(app.isDisplayingJobs)
 
                 // Tap Dashboard 'Back' Button
-                app.navigationBars["Jobs"].buttons["Dashboard"].tap()
+                app.navigationBars[accountName].buttons["Dashboard"].tap()
             }
         }
     }
 
     func testSelectingAJob() {
+        let accountName = "All Inboxes"
         let jobTitle = "Full Stack Dev with Rails Focus"
 
         app.launch()
@@ -135,8 +138,8 @@ class SegueUITests: XCTestCase {
         app.runWithSupportedOrientations {
             startAndEndOnDashboard {
                 // Tap first cell in 'Dashboard' table
-                app.tables["dashboardTableView"].cells.staticTexts["All Inboxes"].tap()
-                app.tables["jobsTableView"].cells.staticTexts[jobTitle].tap()
+                app.tables["dashboardTableView"].cells.staticTexts[accountName].tap()
+                app.cellByLabel(table: app.tables["jobsTableView"], label: jobTitle).tap()
 
                 // iPad should still show list of Jobs
                 if app.isPad() {
@@ -147,9 +150,9 @@ class SegueUITests: XCTestCase {
 
                 // Tap Dashboard 'Back' Button
                 if app.isPhone() {
-                    app.navigationBars[jobTitle].buttons["Jobs"].tap()
+                    app.navigationBars[jobTitle].buttons[accountName].tap()
                 }
-                app.navigationBars["Jobs"].buttons["Dashboard"].tap()
+                app.navigationBars[accountName].buttons["Dashboard"].tap()
 
                 // Previously selected Job should sill be shown on iPad
                 if app.isPad() {
